add minhash skeleton

+++ src/ByteBitset.zig
@@ -17,11 +17,11 @@ pub const ByteBitset = extern struct {
         return self.bits[0] == 0 and self.bits[1] == 0 and self.bits[2] == 0 and self.bits[3] == 0;
     }
 
-    pub fn count(self: *const ByteBitset) u8 {
-        return @as(u8, @popCount(u64, self.bits[0]))
-             + @as(u8, @popCount(u64, self.bits[1]))
-             + @as(u8, @popCount(u64, self.bits[2]))
-             + @as(u8, @popCount(u64, self.bits[3]));
+    pub fn count(self: *const ByteBitset) u16 {
+        return @as(u16, @popCount(u64, self.bits[0]))
+             + @as(u16, @popCount(u64, self.bits[1]))
+             + @as(u16, @popCount(u64, self.bits[2]))
+             + @as(u16, @popCount(u64, self.bits[3]));
     }
 
     pub fn set(self: *ByteBitset, index: u8) void {
